restore lazy loading properly
